bootstrap: Set mailname properly (no quotes!)
Otherwise this breaks sending of mail that self-determines our
hostname.

Also, fix main_mailer_type to be correct (which means we also
need relayhost).

This is kinda silly, because we throw out the generated postfix
conf and rewrite it afterwards... but leaves us in a place where
we could use the debian provided config in the future (if we want).